[network] rename NetworkAsyncService to NetworkServiceRef
