+++ network/src/lib.rs
@@ -7,23 +7,23 @@ extern crate log;
 extern crate prometheus;
 
 pub mod helper;
-pub mod net;
-#[cfg(test)]
-mod net_test;
-pub mod network;
 mod network_metrics;
 mod service;
+pub mod service_ref;
+#[cfg(test)]
+mod tests;
+pub mod worker;
 
-pub use network::PeerMsgBroadcasterService;
 pub use network_api::messages::*;
+pub use service_ref::PeerMsgBroadcasterService;
 
 pub use helper::get_unix_ts;
 
-pub use net::build_network_worker;
-pub use network::NetworkAsyncService;
 use network_p2p::PeerId;
 pub use service::NetworkActorService;
+pub use service_ref::NetworkServiceRef;
 use std::borrow::Cow;
+pub use worker::build_network_worker;
 
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct NetworkMessage {
